+++ .github/CHANGELOG.md
@@ -2,8 +2,9 @@
 
 Update <small>_ July 2024</small>
 
-- fix: corrected typo in the embed for simbrief-data support-request (16/07/2024)
-- docs: updated the Ground Rules on the Contributing guide page (16/07/2024)
+- fix: corrected typo in the embed for simbrief-data support-request (27/07/2024)
+- docs: updated the Ground Rules on the Contributing guide page (27/07/2024)
+- feat: add optional description to `/locate` panels. (27/07/2024)
 
 Update <small>_ May 2024</small>